refactor wallet modal components
