+++ src/components/molecules/WalletItem/walletItem.molecule.types.tsx
@@ -1,0 +1,10 @@
+import {EConnectorType} from "../../../shared/types/web3.types";
+
+interface ConnectFunction {
+  (): void
+}
+
+export interface IWalletItem {
+  connectorType: EConnectorType
+  onClick: ConnectFunction 
+}
