+++ src/components/organisms/Modals/ModalConnectWalletV2/modalConnectWalletV2.organism.tsx
@@ -1,11 +1,19 @@
-import React from "react";
+import React, {useEffect} from "react";
 import {useSelector, useDispatch} from "react-redux";
 import { Modal } from "react-bootstrap";
 import { EModalName } from "../../../../redux/ui/ui.redux.types";
 import { RootState } from "../../../../redux/redux.types";
 import { toggleModal } from "../../../../redux/ui/ui.redux.actions";
-import WalletList from "../../WalletList/walletList.organism";
+// import WalletList from "../../WalletList/walletList.organism";
 import styles from "./modalConnectWalletV2.organism.module.scss";
+import {EConnectorType} from "../../../../shared/types/web3.types";
+import useWeb3 from "../../../../hooks/useWeb3";
+import {supportedConnectorList} from "../../../../shared/constants/web3.constants";
+import WalletService from "../../../../shared/services/wallet/wallet.service";
+import {EErrorTypes} from "../../../../shared/types/error.types";
+import {WalletConnectConnector} from "@web3-react/walletconnect-connector";
+import WalletItem from "../../../molecules/WalletItem/walletItem.molecule";
+
 
 const ModalConnectWalletV2: React.FC = () => {
   const name = EModalName.CONNECT_WALLET_V2;
@@ -16,6 +24,39 @@ const ModalConnectWalletV2: React.FC = () => {
     dispatch(toggleModal(name, false));
   };
 
+  const { activate, active, error, connector, mappedError } = useWeb3();
+
+  const connectWalletFnFactory = (type_: EConnectorType) => {
+    return () => {
+      const _connector = WalletService.typeToProvider(type_);
+      activate(_connector).finally(() => closeModal());
+    };
+  };
+
+  useEffect(() => {
+    if (error) {
+      if (connector instanceof WalletConnectConnector) {
+        if (mappedError === EErrorTypes.UNSUPPORTED_CHAIN) {
+          closeModal();
+        } else if (connector.walletConnectProvider?.wc?.uri) {
+          connector.walletConnectProvider = undefined; // Makes the dialog be opennable after cancellation
+        }
+      }
+    }
+  }, [error]);
+
+  useEffect(() => {
+    if (active) {
+      dispatch(toggleModal(EModalName.CONNECT_WALLET_V2, false));
+    }
+  }, [active]);
+
+  const renderSupportedWallets = () => (
+    supportedConnectorList.map((type_: EConnectorType) => (
+      <WalletItem connectorType={type_} onClick={connectWalletFnFactory(type_)} /> 
+    )
+  ));
+
   return (
     <Modal 
       show={visible} 
@@ -25,7 +66,9 @@ const ModalConnectWalletV2: React.FC = () => {
       centered
     >
       <Modal.Body className={styles.selectWalletModalBody}>
-        <WalletList /> 
+        <ul className={styles.walletList}>
+          {renderSupportedWallets()} 
+        </ul>
       </Modal.Body>
     </Modal>
   );
